fix : 충돌 코드 수정

+++ src/main/java/ourtine/web/dto/response/UserProfileDto.java
@@ -17,7 +17,7 @@ public class UserProfileDto {
     private String goal;
     //private List<Category> userCategoryList;
     private List<String> userCategoryList;
-    private BigDecimal participationRate;
+    private Integer participationRate;
     private long habitCount;
     private boolean isFollowing;
     private long followerCount;
